Use `df.to_json` vs. `sanitize_dataframe`
