--- jupyterlab_table/utils.py
@@ -1,86 +1,0 @@
-import cgi
-import codecs
-import collections
-import os.path
-import json
-import pandas as pd
-
-
-def nested_update(d, u):
-    """Update nested dictionary d (in-place) with keys from u."""
-    for k, v in u.items():
-        if isinstance(v, collections.Mapping):
-            d[k] = nested_update(d.get(k, {}), v)
-        else:
-            d[k] = v
-    return d
-
-
-def abs_path(path):
-    """Make path absolute."""
-    return os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
-        path)
-
-
-def get_content(path):
-    """Get content of file."""
-    with codecs.open(abs_path(path), encoding='utf-8') as f:
-        return f.read()
-
-
-def escape(string):
-    """Escape the string."""
-    return cgi.escape(string, quote=True)
-
-
-def sanitize_dataframe(df):
-    """Sanitize a DataFrame to prepare it for serialization.
-
-    * Make a copy
-    * Raise ValueError if it has a hierarchical index.
-    * Convert categoricals to strings.
-    * Convert np.int dtypes to Python int objects
-    * Convert floats to objects and replace NaNs by None.
-    * Convert DateTime dtypes into appropriate string representations
-    """
-    import pandas as pd
-    import numpy as np
-
-    df = df.copy()
-
-    if isinstance(df.index, pd.core.index.MultiIndex):
-        raise ValueError('Hierarchical indices not supported')
-    if isinstance(df.columns, pd.core.index.MultiIndex):
-        raise ValueError('Hierarchical indices not supported')
-
-    for col_name, dtype in df.dtypes.iteritems():
-        if str(dtype) == 'category':
-            # XXXX: work around bug in to_json for categorical types
-            # https://github.com/pydata/pandas/issues/10778
-            df[col_name] = df[col_name].astype(str)
-        elif np.issubdtype(dtype, np.integer):
-            # convert integers to objects; np.int is not JSON serializable
-            df[col_name] = df[col_name].astype(object)
-        elif np.issubdtype(dtype, np.floating):
-            # For floats, convert nan->None: np.float is not JSON serializable
-            col = df[col_name].astype(object)
-            df[col_name] = col.where(col.notnull(), None)
-        elif str(dtype).startswith('datetime'):
-            # Convert datetimes to strings
-            # astype(str) will choose the appropriate resolution
-            df[col_name] = df[col_name].astype(str).replace('NaT', '')
-    return df
-
-
-def prepare_data(data=None, schema=None):
-    """Prepare JSONTable data from Pandas DataFrame."""
-    
-    if isinstance(data, pd.DataFrame):
-        data = sanitize_dataframe(data)
-        data = data.to_json(orient='table')
-        return json.loads(data)
-    return {
-        'data': data,
-        'schema': schema
-    }
